% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{alpha}
\alias{alpha}
\title{alpha}
\usage{
alpha(col, alpha)
}
\arguments{
\item{col}{RGB color}

\item{alpha}{value of alpha}
}
\description{
Takes provided colors and gives them the specified alpha (ie transparency) value
}
