% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{bam.cov.exome}
\alias{bam.cov.exome}
\title{Get coverage as GRanges from BAM using exons as tiles}
\usage{
bam.cov.exome(bam.file, chunksize = 1e+06, min.mapq = 1,
  verbose = TRUE, max.tlen = 10000, st.flag = "-f 0x02 -F 0x10",
  fragments = TRUE, do.gc = FALSE)
}
\arguments{
\item{bam.file}{string Input BAM file}

\item{chunksize}{integer Size of window (default = 1e5)}

\item{min.mapq}{integer Minimim map quality reads to consider for counts (default = 30)}

\item{verbose}{boolean Flag to increase vebosity (default = TRUE)}

\item{max.tlen}{integer Maximum paired-read insert size to consider (default = 1e4)}

\item{st.flag}{string Samtools flag to filter reads on (default = '-f 0x02 -F 0x10')}

\item{fragments}{boolean flag (default = FALSE) detremining whether to compute fragment (i.e. proper pair footprint aka insert) density or read density}

\item{do.gc}{boolean Flag to execute garbage collection via 'gc()' (default = FALSE)}
}
\value{
GRanges of exon tiles across seqlengths of bam.file with meta data field $counts specifying fragment counts centered (default = TRUE)
in the given bin.
}
\description{
Quick way to get tiled coverage via piping to samtools (e.g. ~10 CPU-hours for 100bp tiles, 5e8 read pairs)

Gets coverage for exons, pulling "chunksize" records at a time and incrementing bin
}
\author{
Trent Walradt
}
