% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{coco}
\alias{coco}
\title{Coverage correction}
\usage{
coco(cov, base = max(10, 1e5/max(width(cov))), fields = c("gc", "map"), iterative = TRUE,
presegment = TRUE, min.segwidth = 5e6, verbose = TRUE,
FUN = NULL, mc.cores = 1, exome = TRUE, imageroot = NULL)
}
\arguments{
\item{cov}{GRanges constant with GRanges of coverage samples with (by default) fields $reads, $map, $gc}

\item{base}{integer Scale multiplier}

\item{fields}{character vector fields of gc to use as covariates}

\item{iterative}{boolean whether to iterate}

\item{presegment}{boolean whether to presegment}

\item{min.segwidth}{integer when presegmenting, minimum segment width}

\item{verbose}{boolean Wether to print log to console}

\item{FUN}{function with which to correct coverage (by default loess
correction modified from HMMcopy that takes in granges with fields
$reads and other fields specified in "fields"}

\item{...}{additional args to FUN}

\item{mc.cores}{integer Number of cores to use}

\item{exome}{boolean If TRUE, collapse by 1e5 for presegmentation}

\item{imageroot}{String, optional file root to which to dump images of correction}
}
\description{
Given gc and mappability coverage correction, yields corrected
read counts
}
\author{
Trent Walradt
}
