% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{fragCounter}
\alias{fragCounter}
\title{fragCounter}
\usage{
fragCounter(
  bam,
  skeleton,
  cov = NULL,
  midpoint = TRUE,
  window = 200,
  gc.rds.dir,
  map.rds.dir,
  minmapq = 20,
  reference = NULL,
  paired = TRUE,
  outdir = NULL,
  exome = FALSE,
  use.skel = FALSE,
  chr.sub = TRUE,
  st.flag = fragCounter::ST.FLAG,
  min.tlen = 0,
  max.tlen = 10000,
  correct_for_bias = TRUE
)
}
\arguments{
\item{bam}{string path to .bam or .cram file}

\item{skeleton}{string Input data.table with intervals for which there is coverage data}

\item{cov}{string path to existing coverage rds or bedgraph}

\item{midpoint}{boolean If TRUE only count midpoint if FALSE then count bin footprint of every fragment interval}

\item{window}{integer window / bin size}

\item{gc.rds.dir}{string for tiles of width W, will look here for a file named gc{W}.rds in this directory}

\item{map.rds.dir}{string for tiles of width W will look here for a file named map{W}.rds in this directory}

\item{minmapq}{double Minimal map quality}

\item{reference}{reference file (recommended for CRAM, [NULL]}

\item{paired}{boolean wether or not paired}

\item{outdir}{string Directory to dump output into}

\item{exome}{boolean If TRUE, perform correction using exons as bins instead of fixed size}

\item{use.skel}{boolean flag If false then default exome skeleton from gencode is used, if TRUE, user defined skeleton is usde}

\item{chr.sub}{boolean if TRUE, remove 'chr' prefix on seqnames. default TRUE}

\item{min.tlen}{numeric Minimum template length (insert size)}

\item{max.tlen}{numeric Maximum template length (insert size)}
}
\description{
Runs entire fragCounter pipeline
}
\author{
Marcin Imielinski
}
