% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{make.blacklist}
\alias{make.blacklist}
\title{Make a list of regions of the genome in which there is coverage in the normal samples}
\usage{
make.blacklist(pairs, cutoff = 0.9)
}
\arguments{
\item{pairs}{string Input data.table with columns 'pair' and 'bam' corresponding to the sample name and bath to normal bam}

\item{cutoff}{double Threshold for what percentage of normal samples must have reads for the region to be included}
}
\value{
Data.table of regions to include when calculating corrected reads
}
\description{
Look at all available normal samples and find regions where some percentage of all normal samples have at least 1 read
}
\author{
Trent Walradt
}
