% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{multicoco}
\alias{multicoco}
\title{Multi-scale coverage correction}
\usage{
multicoco(cov, numlevs = 1, fields = c("gc", "map"), iterative = TRUE,
presegment = TRUE, min.segwidth = 5e6, mono = FALSE, verbose = TRUE,
FUN = NULL, mc.cores = 1)
}
\arguments{
\item{cov}{GRanges constant with GRanges of coverage samples with (by default) fields $reads, $map, $gc}

\item{numlevs}{integer numbers of scales at which to correct}

\item{base}{integer Scale multiplier}

\item{fields}{character vector fields of gc to use as covariates}

\item{iterative}{boolean whether to iterate}

\item{presegment}{boolean whether to presegment}

\item{min.segwidth}{integer when presegmenting, minimum segment width}

\item{mono}{boolean Wether to only do single iteration at finest scale}

\item{verbose}{boolean Wether to print log to console}

\item{FUN}{function with which to correct coverage (by default loess
correction modified from HMMcopy that takes in granges with fields
$reads and other fields specified in "fields"}

\item{...}{additional args to FUN}

\item{mc.cores}{integer Number of cores to use}
}
\description{
Given gc and mappability coverage correction at k "nested" scales finds the coverage
assignment at the finest scale that yields the best correction at every scale
}
\author{
Marcin Imielinski
}
