% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\docType{data}
\name{ST.FLAG}
\alias{ST.FLAG}
\title{Samtools flags}
\format{
An object of class \code{character} of length 1.
}
\usage{
ST.FLAG
}
\description{
Samtools flags for fragCounter bam query
-F exclude ANY bits set
3868 sets bits
4 - Read unmapped
8 - Mate unmapped
16 - Read reverse strand
256 - Not primary alignment (secondary alignment)
512 - Read fails platform/vendor quality checks
1024 - Read is PCR or optical duplicate
2048 - Supplementary alignment (chimeric, non-contiguous alignment)
-f requires read have ALL bits set
2 - Read mapped in proper pair
}
\keyword{datasets}
