% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{fragCounter}
\alias{fragCounter}
\title{fragCounter}
\usage{
fragCounter(bam, cov = NULL, midpoint = FALSE, window = 200, gc.rds.dir,
  map.rds.dir, minmapq = 1, paired = TRUE, outdir = NULL)
}
\arguments{
\item{bam}{string path to .bam file}

\item{cov}{string path to existing coverage rds or bedgraph}

\item{midpoint}{boolean If TRUE only count midpoint if FALSE then count bin footprint of every fragment interval}

\item{window}{integer window / bin size}

\item{gc.rds.dir}{string for tiles of width W, will look here for a file named gc{W}.rds in this directory}

\item{map.rds.dir}{string for tiles of width W will look here for a file named map{W}.rds in this directory}

\item{minmapq}{double Minimal map quality}

\item{paired}{boolean wether or not paired}

\item{outdir}{string Directory to dump output into}
}
\description{
Runs entire fragCounter pipeline
}
\author{
Marcin Imielinski
}
