% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{GC.fun}
\alias{GC.fun}
\title{GC content calculator}
\usage{
GC.fun(win.size = 200, twobitURL = "~/DB/UCSC/hg19.2bit",
  twobit.win = 1000, mc.cores = 10)
}
\arguments{
\item{win.size}{Window size, in basepairs, to calculate GC content for (should match window
size of your coverage file}

\item{twobitURL}{URL to twobit genome file. Default is hg19 from UCSC}

\item{twobit.win}{How many windows of the twobit file to load into memory on each core}

\item{mc.cores}{How many cores to use}
}
\description{
Calculates GC content across the genome for a given sized window
}
\author{
Marcin Imielinski
}
