% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragCounter.R
\name{PrepareCov}
\alias{PrepareCov}
\title{PrepareCov}
\usage{
PrepareCov(bam, skeleton, cov = NULL, midpoint = TRUE, window = 200,
  minmapq = 1, paired = TRUE, outdir = NULL, exome = FALSE,
  use.skel = FALSE)
}
\arguments{
\item{bam}{path to .bam file}

\item{skeleton}{string Input data.table with intervals for which there is coverage data}

\item{cov}{Path to existing coverage rds or bedgraph}

\item{midpoint}{If TRUE only count midpoint if FALSE then count bin footprint of every fragment interval}

\item{window}{window / bin size}

\item{minmapq}{Minimal map quality}

\item{paired}{wether or not paired}

\item{outdir}{Directory to dump output into}

\item{exome}{boolean If TRUE, use bam.cov.exome to calculate coverage}

\item{use.skel}{boolean flag If false then default exome skeleton from gencode is used, if TRUE, user defined skeleton is used}
}
\description{
Load BAM or coverage file and prepare for use in correctcov_stub
}
\author{
Marcin Imielinski

Trent Walradt
}
